+++ lib/model/alarm_settings.dart
@@ -1,27 +1,31 @@
 import 'package:alarm/model/notification_settings.dart';
+import 'package:alarm/model/volume_settings.dart';
 import 'package:alarm/src/generated/platform_bindings.g.dart';
-import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+part 'alarm_settings.g.dart';
 
 /// [AlarmSettings] is a model that contains all the settings to customize
 /// and set an alarm.
-@immutable
+@JsonSerializable()
 class AlarmSettings {
   /// Constructs an instance of `AlarmSettings`.
   const AlarmSettings({
     required this.id,
     required this.dateTime,
     required this.assetAudioPath,
+    required this.volumeSettings,
     required this.notificationSettings,
     this.loopAudio = true,
     this.vibrate = true,
-    this.volume,
-    this.volumeEnforced = false,
-    this.fadeDuration = 0.0,
-    this.fadeStopTimes = const [],
-    this.fadeStopVolumes = const [],
     this.warningNotificationOnKill = true,
     this.androidFullScreenIntent = true,
   });
+
+  // TODO(orkun1675): Ensure compatibility with older plugin versions.
+  /// Constructs an `AlarmSettings` instance from the given JSON data.
+  factory AlarmSettings.fromJson(Map<String, dynamic> json) =>
+      _$AlarmSettingsFromJson(json);
 
   /// Converts from wire datatype.
   AlarmSettings.fromWire(AlarmSettingsWire wire)
@@ -29,63 +33,13 @@ class AlarmSettings {
         dateTime =
             DateTime.fromMillisecondsSinceEpoch(wire.millisecondsSinceEpoch),
         assetAudioPath = wire.assetAudioPath,
+        volumeSettings = VolumeSettings.fromWire(wire.volumeSettings),
         notificationSettings =
             NotificationSettings.fromWire(wire.notificationSettings),
         loopAudio = wire.loopAudio,
         vibrate = wire.vibrate,
-        volume = wire.volume,
-        volumeEnforced = wire.volumeEnforced,
-        fadeDuration = wire.fadeDuration,
         warningNotificationOnKill = wire.warningNotificationOnKill,
         androidFullScreenIntent = wire.androidFullScreenIntent;
-
-  /// Constructs an `AlarmSettings` instance from the given JSON data.
-  factory AlarmSettings.fromJson(Map<String, dynamic> json) {
-    NotificationSettings notificationSettings;
-
-    // Ensure compatibility with plugin versions below 4.0.0.
-    if (json.containsKey('notificationSettings') &&
-        json['notificationSettings'] != null) {
-      notificationSettings = NotificationSettings.fromJson(
-        json['notificationSettings'] as Map<String, dynamic>,
-      );
-    } else {
-      final notificationTitle = json['notificationTitle'] as String? ?? '';
-      final notificationBody = json['notificationBody'] as String? ?? '';
-
-      notificationSettings = NotificationSettings(
-        title: notificationTitle,
-        body: notificationBody,
-      );
-    }
-
-    final warningNotificationOnKill =
-        json.containsKey('warningNotificationOnKill')
-            ? json['warningNotificationOnKill'] as bool
-            : json['enableNotificationOnKill'] as bool? ?? true;
-
-    return AlarmSettings(
-      id: json['id'] as int,
-      dateTime: DateTime.fromMicrosecondsSinceEpoch(json['dateTime'] as int),
-      assetAudioPath: json['assetAudioPath'] as String,
-      notificationSettings: notificationSettings,
-      loopAudio: json['loopAudio'] as bool? ?? true,
-      vibrate: json['vibrate'] as bool? ?? true,
-      volume: json['volume'] as double?,
-      volumeEnforced: json['volumeEnforced'] as bool? ?? false,
-      fadeDuration: json['fadeDuration'] as double? ?? 0.0,
-      fadeStopTimes: (json['fadeStopTimes'] as List<dynamic>?)
-              ?.map((e) => e as double)
-              .toList(growable: false) ??
-          [],
-      fadeStopVolumes: (json['fadeStopVolumes'] as List<dynamic>?)
-              ?.map((e) => e as double)
-              .toList(growable: false) ??
-          [],
-      warningNotificationOnKill: warningNotificationOnKill,
-      androidFullScreenIntent: json['androidFullScreenIntent'] as bool? ?? true,
-    );
-  }
 
   /// Unique identifier assiocated with the alarm. Cannot be 0 or -1;
   final int id;
@@ -116,6 +70,9 @@ class AlarmSettings {
   /// `android.permission.READ_EXTERNAL_STORAGE`
   final String assetAudioPath;
 
+  /// Settings for the alarm volume.
+  final VolumeSettings volumeSettings;
+
   /// Settings for the notification.
   final NotificationSettings notificationSettings;
 
@@ -127,55 +84,6 @@ class AlarmSettings {
   ///
   /// If [loopAudio] is set to false, vibrations will stop when audio ends.
   final bool vibrate;
-
-  /// Specifies the system volume level to be set at the designated [dateTime].
-  ///
-  /// Accepts a value between 0 (mute) and 1 (maximum volume).
-  /// When the alarm is triggered at [dateTime], the system volume adjusts to
-  /// this specified level. Upon stopping the alarm, the system volume reverts
-  /// to its prior setting.
-  ///
-  /// If left unspecified or set to `null`, the current system volume
-  /// at the time of the alarm will be used.
-  /// Defaults to `null`.
-  final double? volume;
-
-  /// If true, the alarm volume is enforced, automatically resetting to the
-  /// original alarm [volume] if the user attempts to adjust it.
-  /// This prevents the user from lowering the alarm volume.
-  /// Won't work if app is killed.
-  ///
-  /// Defaults to false.
-  final bool volumeEnforced;
-
-  /// Duration, in seconds, over which to fade the alarm ringtone.
-  /// Set to 0.0 by default, which means no fade.
-  final double fadeDuration;
-
-  /// Together with [fadeStopVolumes] controls how the alarm volume will fade
-  /// over time.
-  ///
-  /// Set to empty list by default, which means no fade.
-  ///
-  /// Must have positive values and the same length as [fadeStopVolumes].
-  ///
-  /// Example:
-  ///    fadeStopTimes = [0, 10, 20]
-  ///    fadeStopVolumes = [0, 0.5, 1.0]
-  /// The alarm will begin silent, fade to 50% of max volume by 10 seconds,
-  /// and fade to max volume by 20 seconds.
-  final List<double> fadeStopTimes;
-
-  /// Together with [fadeStopTimes] controls how the alarm volume will fade
-  /// over time.
-  ///
-  /// Set to empty list by default, which means no fade.
-  ///
-  /// Must have values that are in [0, 1] and have the same length as
-  /// [fadeStopTimes].
-  ///
-  /// See [fadeStopTimes] for more information.
-  final List<double> fadeStopVolumes;
 
   /// Whether to show a warning notification when application is killed by user.
   ///
@@ -196,44 +104,21 @@ class AlarmSettings {
   /// package.
   final bool androidFullScreenIntent;
 
+  /// Converts the `AlarmSettings` instance to a JSON object.
+  Map<String, dynamic> toJson() => _$AlarmSettingsToJson(this);
+
   /// Converts to wire datatype which is used for host platform communication.
   AlarmSettingsWire toWire() => AlarmSettingsWire(
         id: id,
         millisecondsSinceEpoch: dateTime.millisecondsSinceEpoch,
         assetAudioPath: assetAudioPath,
+        volumeSettings: volumeSettings.toWire(),
         notificationSettings: notificationSettings.toWire(),
         loopAudio: loopAudio,
         vibrate: vibrate,
-        volume: volume,
-        volumeEnforced: volumeEnforced,
-        fadeDuration: fadeDuration,
         warningNotificationOnKill: warningNotificationOnKill,
         androidFullScreenIntent: androidFullScreenIntent,
       );
-
-  /// Returns a hash code for this `AlarmSettings` instance using
-  /// Jenkins hash function.
-  @override
-  int get hashCode {
-    var hash = 0;
-
-    hash = hash ^ id.hashCode;
-    hash = hash ^ dateTime.hashCode;
-    hash = hash ^ assetAudioPath.hashCode;
-    hash = hash ^ notificationSettings.hashCode;
-    hash = hash ^ loopAudio.hashCode;
-    hash = hash ^ vibrate.hashCode;
-    hash = hash ^ volume.hashCode;
-    hash = hash ^ volumeEnforced.hashCode;
-    hash = hash ^ fadeDuration.hashCode;
-    hash = hash ^ Object.hashAll(fadeStopTimes);
-    hash = hash ^ Object.hashAll(fadeStopVolumes);
-    hash = hash ^ warningNotificationOnKill.hashCode;
-    hash = hash ^ androidFullScreenIntent.hashCode;
-    hash = hash & 0x3fffffff;
-
-    return hash;
-  }
 
   /// Creates a copy of `AlarmSettings` but with the given fields replaced with
   /// the new values.
@@ -241,6 +126,7 @@ class AlarmSettings {
     int? id,
     DateTime? dateTime,
     String? assetAudioPath,
+    VolumeSettings? volumeSettings,
     NotificationSettings? notificationSettings,
     bool? loopAudio,
     bool? vibrate,
@@ -258,64 +144,14 @@ class AlarmSettings {
       id: id ?? this.id,
       dateTime: dateTime ?? this.dateTime,
       assetAudioPath: assetAudioPath ?? this.assetAudioPath,
+      volumeSettings: volumeSettings ?? this.volumeSettings,
       notificationSettings: notificationSettings ?? this.notificationSettings,
       loopAudio: loopAudio ?? this.loopAudio,
       vibrate: vibrate ?? this.vibrate,
-      volume: volume ?? this.volume,
-      volumeEnforced: volumeEnforced ?? this.volumeEnforced,
-      fadeDuration: fadeDuration ?? this.fadeDuration,
-      fadeStopTimes: fadeStopTimes ?? this.fadeStopTimes,
-      fadeStopVolumes: fadeStopVolumes ?? this.fadeStopVolumes,
       warningNotificationOnKill:
           warningNotificationOnKill ?? this.warningNotificationOnKill,
       androidFullScreenIntent:
           androidFullScreenIntent ?? this.androidFullScreenIntent,
     );
   }
-
-  /// Converts this `AlarmSettings` instance to JSON data.
-  Map<String, dynamic> toJson() => {
-        'id': id,
-        'dateTime': dateTime.microsecondsSinceEpoch,
-        'assetAudioPath': assetAudioPath,
-        'notificationSettings': notificationSettings.toJson(),
-        'loopAudio': loopAudio,
-        'vibrate': vibrate,
-        'volume': volume,
-        'volumeEnforced': volumeEnforced,
-        'fadeDuration': fadeDuration,
-        'fadeStopTimes': fadeStopTimes,
-        'fadeStopVolumes': fadeStopVolumes,
-        'warningNotificationOnKill': warningNotificationOnKill,
-        'androidFullScreenIntent': androidFullScreenIntent,
-      };
-
-  /// Returns all the properties of `AlarmSettings` for debug purposes.
-  @override
-  String toString() {
-    final json = toJson();
-    json['dateTime'] =
-        DateTime.fromMicrosecondsSinceEpoch(json['dateTime'] as int);
-
-    return 'AlarmSettings: $json';
-  }
-
-  /// Compares two AlarmSettings.
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is AlarmSettings &&
-          runtimeType == other.runtimeType &&
-          id == other.id &&
-          dateTime == other.dateTime &&
-          assetAudioPath == other.assetAudioPath &&
-          notificationSettings == other.notificationSettings &&
-          loopAudio == other.loopAudio &&
-          vibrate == other.vibrate &&
-          volume == other.volume &&
-          fadeDuration == other.fadeDuration &&
-          listEquals(fadeStopTimes, other.fadeStopTimes) &&
-          listEquals(fadeStopVolumes, other.fadeStopVolumes) &&
-          warningNotificationOnKill == other.warningNotificationOnKill &&
-          androidFullScreenIntent == other.androidFullScreenIntent;
 }